[ #2083 ] Moved the internal tests to a Cabal test-suite (third part).
