+++ test/InternalTests.hs
@@ -4,39 +4,39 @@ module Main ( main ) where
 
 import InternalTests.Helpers
 
-import InternalTests.Compiler.MAlonzo.Encode              as CompEnco     ( tests )
-import InternalTests.Interaction.Highlighting.Emacs       as InteEmac     ( tests )
-import InternalTests.Interaction.Highlighting.Generate    as InteGene     ( tests )
-import InternalTests.Interaction.Highlighting.Precise     as IntePrec     ( tests )
-import InternalTests.Interaction.Highlighting.Range       as InteRang     ( tests )
-import InternalTests.Interaction.Options                  as InteOpti     ( tests )
-import InternalTests.Syntax.Parser.Parser                 as SyntPars     ( tests )
-import Agda.Syntax.Position                               as SyntPosi     ( tests )
-import Agda.Termination.CallGraph                         as TermCall     ( tests )
-import Agda.Termination.CallMatrix                        as TermCM       ( tests )
-import Agda.Termination.Order                             as TermOrd      ( tests )
-import Agda.Termination.Semiring                          as TermRing     ( tests )
-import Agda.Termination.SparseMatrix                      as TermSparse   ( tests )
-import Agda.Termination.Termination                       as TermTerm     ( tests )
-import InternalTests.TypeChecking.Free                    as Free         ( tests )
-import Agda.TypeChecking.Irrelevance                      as Irrel        ( tests )
-import Agda.TypeChecking.Positivity.Tests                 as Positivity   ( tests )
-import Agda.TypeChecking.Positivity.Occurrence            as Occurrence   ( tests )
-import Agda.TypeChecking.Tests                            as TypeChck     ( tests )
-import Agda.TypeChecking.SizedTypes.Tests                 as SizedTypes   ( tests )
-import Agda.Utils.Bag                                     as UtilBag      ( tests )
-import Agda.Utils.BiMap                                   as UtilBiMap    ( tests )
-import Agda.Utils.Cluster                                 as UtilClust    ( tests )
-import Agda.Utils.Either                                  as UtilEith     ( tests )
-import Agda.Utils.Favorites                               as UtilFav      ( tests )
-import Agda.Utils.FileName                                as UtilFile     ( tests )
-import Agda.Utils.Graph.AdjacencyMap.Unidirectional.Tests as UtilGraphUni ( tests )
-import Agda.Utils.List                                    as UtilList     ( tests )
-import Agda.Utils.ListT.Tests                             as UtilListT    ( tests )
-import Agda.Utils.PartialOrd                              as UtilPOrd     ( tests )
-import Agda.Utils.Permutation.Tests                       as UtilPerm     ( tests )
-import Agda.Utils.Trie                                    as UtilTrie     ( tests )
-import Agda.Utils.Warshall                                as UtilWarsh    ( tests )
+import InternalTests.Compiler.MAlonzo.Encode                  as CompEnco     ( tests )
+import InternalTests.Interaction.Highlighting.Emacs           as InteEmac     ( tests )
+import InternalTests.Interaction.Highlighting.Generate        as InteGene     ( tests )
+import InternalTests.Interaction.Highlighting.Precise         as IntePrec     ( tests )
+import InternalTests.Interaction.Highlighting.Range           as InteRang     ( tests )
+import InternalTests.Interaction.Options                      as InteOpti     ( tests )
+import InternalTests.Syntax.Parser.Parser                     as SyntPars     ( tests )
+import Agda.Syntax.Position                                   as SyntPosi     ( tests )
+import Agda.Termination.CallGraph                             as TermCall     ( tests )
+import Agda.Termination.CallMatrix                            as TermCM       ( tests )
+import Agda.Termination.Order                                 as TermOrd      ( tests )
+import Agda.Termination.Semiring                              as TermRing     ( tests )
+import Agda.Termination.SparseMatrix                          as TermSparse   ( tests )
+import Agda.Termination.Termination                           as TermTerm     ( tests )
+import InternalTests.TypeChecking.Free                        as Free         ( tests )
+import InternalTests.TypeChecking.Irrelevance                 as Irrel        ( tests )
+import InternalTests.TypeChecking.Positivity                  as Positivity   ( tests )
+import Agda.TypeChecking.Positivity.Occurrence                as Occurrence   ( tests )
+import Agda.TypeChecking.Tests                                as TypeChck     ( tests )
+import InternalTests.TypeChecking.SizedTypes                  as SizedTypes   ( tests )
+import Agda.Utils.Bag                                         as UtilBag      ( tests )
+import Agda.Utils.BiMap                                       as UtilBiMap    ( tests )
+import Agda.Utils.Cluster                                     as UtilClust    ( tests )
+import Agda.Utils.Either                                      as UtilEith     ( tests )
+import Agda.Utils.Favorites                                   as UtilFav      ( tests )
+import Agda.Utils.FileName                                    as UtilFile     ( tests )
+import InternalTests.Utils.Graph.AdjacencyMap.Unidirectional  as UtilGraphUni ( tests )
+import Agda.Utils.List                                        as UtilList     ( tests )
+import Agda.Utils.ListT.Tests                                 as UtilListT    ( tests )
+import Agda.Utils.PartialOrd                                  as UtilPOrd     ( tests )
+import InternalTests.Utils.Permutation                        as UtilPerm     ( tests )
+import InternalTests.Utils.Trie                               as UtilTrie     ( tests )
+import InternalTests.Utils.Warshall                           as UtilWarsh    ( tests )
 
 main :: IO Bool
 main = runTests "QuickCheck test suite:"
